# Extract doc comments