Closes #10682

Opening as a draft for now.  

todo: decl's and fields