@I-mikan-I 2022-05-12
> Hi, thank you for the patch. I would like to request the following things:
> 
>     * clean up the git commit history
> 
>     * rebase against latest master, address the conflicts
> 
>     * avoid the std lib changes which should be unnecessary now since the flaky test has been disabled
> 
>     * provide some kind of perf measurement to share whether this had any noticeable impact

Thanks, about the last point:
Do you think benchmarking the entire compilation process makes sense in this case? Or should I constrain benchmarks to some smaller unit (maybe there's an obvious one)? Perhaps the impact wouldn't be noticeable otherwise.