@andrewrk 2022-05-12
Here's one possible strategy:

```
$ ./zig build -p stage2-release -Denable-llvm -Dskip-install-lib-files -Drelease
$ perf stat -e branch-instructions -e branch-misses -e cache-misses -e cache-references -e cpu-cycles -e instructions ./stage2-release/bin/zig build -p stage3 -Dskip-install-lib-files
```

Then compare this against master branch.