@I-mikan-I 2022-05-14
After changes:
```
 Performance counter stats for './stage2-release/bin/zig build -p stage3 -Dskip-install-lib-files':

    40,390,993,583      branch-instructions:u                                         (83.38%)
     1,412,832,604      branch-misses:u           #    3.50% of all branches          (83.36%)
       716,631,229      cache-misses:u            #    7.516 % of all cache refs      (83.24%)
     9,535,235,940      cache-references:u                                            (83.32%)
   167,487,191,242      cpu-cycles:u                                                  (66.72%)
   215,611,750,504      instructions:u            #    1.29  insn per cycle           (83.40%)

      36.498526829 seconds time elapsed

      35.546125000 seconds user
       1.035131000 seconds sys
```
Master:
```
    40,355,037,914      branch-instructions:u                                         (83.37%)
     1,410,952,944      branch-misses:u           #    3.50% of all branches          (83.39%)
       685,439,676      cache-misses:u            #    7.303 % of all cache refs      (83.28%)
     9,385,122,977      cache-references:u                                            (83.28%)
   166,070,827,653      cpu-cycles:u                                                  (66.70%)
   215,319,747,032      instructions:u            #    1.30  insn per cycle           (83.35%)

      36.359394807 seconds time elapsed

      35.454425000 seconds user
       0.867833000 seconds sys
```

I suppose cache misses did go down slightly, but I'm not sure if there's any scenario where it's noticeable.