@I-mikan-I 2022-05-14
After changes:
```
 Performance counter stats for './stage2-release/bin/zig build -p stage3 -Dskip-install-lib-files':

    35,821,892,069      branch-instructions:u                                         (83.36%)
     1,261,309,893      branch-misses:u           #    3.52% of all branches          (83.37%)
       596,529,490      cache-misses:u            #    7.062 % of all cache refs      (83.36%)
     8,446,664,652      cache-references:u                                            (83.36%)
   149,459,102,707      cpu-cycles:u                                                  (66.57%)
   191,475,789,729      instructions:u            #    1.28  insn per cycle           (83.36%)

      32.894469784 seconds time elapsed

      31.754609000 seconds user
       0.833348000 seconds sys
```
Master:
```
    35,845,471,420      branch-instructions:u                                         (83.37%)
     1,255,935,732      branch-misses:u           #    3.50% of all branches          (83.39%)
       597,490,049      cache-misses:u            #    7.177 % of all cache refs      (83.35%)
     8,324,773,521      cache-references:u                                            (83.34%)
   147,175,855,756      cpu-cycles:u                                                  (66.59%)
   191,619,988,270      instructions:u            #    1.30  insn per cycle           (83.31%)

      32.326098637 seconds time elapsed

      31.228449000 seconds user
       0.902049000 seconds sys
```
So it doesn't seem like there's a huge impact, the differences could be due to noise. Maybe you would notice a difference if the compiled source code was constructed just to highlight these changes.