@I-mikan-I 2022-05-14
I profiled a few compilation for both master and this branch on a simple executable, which consisted of a large enum and a doc comment per field declaration, with main only referencing all field decls. I got the following results.
With changes:
```
 Performance counter stats for './stage2-release/bin/zig build-exe example.zig':

       541,573,346      branch-instructions:u                                         (84.59%)
        19,729,595      branch-misses:u           #    3.64% of all branches          (84.99%)
         3,212,026      cache-misses:u            #    2.409 % of all cache refs      (83.77%)
       133,351,996      cache-references:u                                            (77.51%)
     2,052,089,774      cpu-cycles:u                                                  (69.53%)
     2,860,628,570      instructions:u            #    1.39  insn per cycle           (84.86%)

       0.484928942 seconds time elapsed

       0.480242000 seconds user
       0.042289000 seconds sys
```
Master:
```
       545,891,898      branch-instructions:u                                         (84.32%)
        19,776,797      branch-misses:u           #    3.62% of all branches          (84.97%)
         3,518,289      cache-misses:u            #    2.778 % of all cache refs      (80.71%)
       126,670,523      cache-references:u                                            (81.89%)
     2,067,844,968      cpu-cycles:u                                                  (68.55%)
     2,861,980,633      instructions:u            #    1.38  insn per cycle           (84.70%)

       0.483269667 seconds time elapsed

       0.446572000 seconds user
       0.075916000 seconds sys
```
There were some deviations between individual runs, the master branch seems to have on average around ~0.3% more cache misses. This is a very unscientific method though.