@andrewrk 2022-05-16
Thank you for all this work. The implementation looks good to me, and I think this PR could be merged. However, given that we can't seem to find any evidence for this being a performance improvement, and this encoding is slightly more complex than master branch, perhaps the best move here will be to close the issue and explain why it is not worth pursuing.

What do you think, @I-mikan-I?