@I-mikan-I 2022-05-17
> Thank you for all this work. The implementation looks good to me, and I think this PR could be merged. However, given that we can't seem to find any evidence for this being a performance improvement, and this encoding is slightly more complex than master branch, perhaps the best move here will be to close the issue and explain why it is not worth pursuing.
> 
> What do you think, @I-mikan-I?

I agree, in this case it would add complexity to the codebase without any direct benefit.
The only reason I see to merge the changes is if the types sema operates on get more such "meta-fields" in the future. It could make sense to save them in a separate array (I imagine eventually there would be a noticeable impact on indexing).
Maybe just something to keep in mind when adding fields in the future.

If you think we'd rather wait on this we can close the PR and issue 🙂.