@andrewrk 2022-05-17
@kristoff-it can you share your opinion on this?

1. Merge the PR, change the way doc comments are accessed from ZIR (potentially simpler for autodocs?)
2. Close the PR, leave it as status quo at least for now